requirements.txt
boto3
botocore
python-dotenv
psycopg2-binary
numpy 
pandas
scikit-learn
pika
imblearn
botocore
statsmodels